optimize the shared lock (#1464)
* optimize the shared lock:
currently the acquire and release of shared lock need to issue a new connection, and
in case client crashed without releasing locks, the lock is dead permanently.
so we introduce persistent connection in LocalSocketComm and it is used in SharedLock
cases, the server will detect if connection is alive and release the lock if the
connection with lock is broken.

* add more unit tests

* fix leak in test_metric_monitor

* the github UT may be slow because of environment resource

* close ThreadPoolExecutor when close server

* use thread instead of ThreadPoolExecute

---------

Co-authored-by: Ma Jie Yue <jieyue.majy@antgroup.com>
Co-authored-by: Tianyi Chen <chentianyi.cty@antfin.com>